test(cli): consolidate scripted Prompts and DockerRunner stubs into a shared helper
Extracts the per-file Prompts stub and DockerRunner baseline literals
(duplicated across six CLI test files since PR #393) into
cli/src/__tests__/command-stubs.ts: one answer-replaying Prompts stub
that records every output channel, three DockerRunner baselines
(dockerReady / dockerDown / dockerDaemonOnly) for spread-and-override
variants, and the shared fetchOk / fetchNever health-check stubs.

The shared stub throws on an unscripted interactive prompt, so the
commands with no interactive prompts (upgrade, lifecycle,
get-sync-token) now fail loudly if a flow unexpectedly asks a question
instead of receiving a silent default answer.

cli/tsconfig.json excludes src/**/__tests__/** so the helper is
type-checked by tsconfig.test.json but never emitted to dist.
