+++ cli/src/__tests__/lifecycle.test.ts
@@ -5,71 +5,13 @@ import { describe, expect, it, onTestFinished } from "vitest"
 
 import type { DockerLogsParams, DockerRunner } from "../docker.js"
 import { runDown, runLogs, runRestart } from "../lifecycle.js"
-import type { Prompts } from "../prompts.js"
-
-const createScriptedPrompts = () => {
-  const errors: string[] = []
-  const logs: string[] = []
-  const outros: string[] = []
-  const spinnerMessages: string[] = []
-
-  const prompts: Prompts = {
-    intro: () => {},
-    outro: (message) => {
-      outros.push(message)
-    },
-    note: () => {},
-    print: () => {},
-    log: (message) => {
-      logs.push(message)
-    },
-    warn: () => {},
-    error: (message) => {
-      errors.push(message)
-    },
-    select: async () => "",
-    multiselect: async () => [],
-    text: async () => "",
-    password: async () => "",
-    confirm: async () => false,
-    spinner: () => ({
-      start: (message) => {
-        spinnerMessages.push(`start: ${message}`)
-      },
-      stop: (message) => {
-        spinnerMessages.push(`stop: ${message}`)
-      },
-    }),
-  }
-
-  return { prompts, errors, logs, outros, spinnerMessages }
-}
-
-const dockerReady: DockerRunner = {
-  isDaemonRunning: () => true,
-  dockerRun: () => true,
-  pullImage: () => true,
-  stopAndRemoveContainer: () => true,
-  containerExists: () => true,
-  streamLogs: async () => 0,
-  runObsidianLogin: () => false,
-}
-
-const dockerDown: DockerRunner = {
-  isDaemonRunning: () => false,
-  dockerRun: () => false,
-  pullImage: () => false,
-  stopAndRemoveContainer: () => false,
-  containerExists: () => false,
-  streamLogs: async () => 1,
-  runObsidianLogin: () => false,
-}
-
-const fetchOk: typeof fetch = async () => new Response(null, { status: 200 })
-
-const fetchNever: typeof fetch = async () => {
-  throw new Error("fetch must not be called")
-}
+import {
+  createScriptedPrompts,
+  dockerDown,
+  dockerReady,
+  fetchNever,
+  fetchOk,
+} from "./command-stubs.js"
 
 // Creates a temp target dir and registers its removal at creation time, so
 // the suite self-cleans even when an assertion throws mid-test.
