+++ cli/src/__tests__/optional-settings.test.ts
@@ -6,87 +6,7 @@ import {
   derivePublicUrlOverride,
   readOptionalValue,
 } from "../optional-settings.js"
-import type { Prompts, SelectOption } from "../prompts.js"
-
-type ScriptedAnswer = string | boolean | string[]
-
-type MultiselectCall = { message: string; options: SelectOption[] }
-type ConfirmCall = { message: string; initialValue: boolean }
-type SelectCall = { message: string; initialValue: string }
-type TextCall = { message: string; defaultValue: string | undefined }
-
-/**
- * A Prompts stub that replays canned answers in order and records what was
- * asked — mirrors the init.test.ts stub, plus per-kind call recording so the
- * chooser's options and each prompt's initial value can be asserted.
- */
-const createScriptedPrompts = (answers: ScriptedAnswer[]) => {
-  const remaining = [...answers]
-  const asked: string[] = []
-  const errors: string[] = []
-  const multiselectCalls: MultiselectCall[] = []
-  const confirmCalls: ConfirmCall[] = []
-  const selectCalls: SelectCall[] = []
-  const textCalls: TextCall[] = []
-
-  const nextAnswer = (message: string): ScriptedAnswer => {
-    asked.push(message)
-    const answer = remaining.shift()
-    if (answer === undefined)
-      throw new Error(`No scripted answer for prompt: ${message}`)
-    return answer
-  }
-
-  const prompts: Prompts = {
-    intro: () => {},
-    outro: () => {},
-    note: () => {},
-    print: () => {},
-    log: () => {},
-    warn: () => {},
-    error: (message) => {
-      errors.push(message)
-    },
-    select: async (message, _options, initialValue) => {
-      selectCalls.push({ message, initialValue })
-      return String(nextAnswer(message))
-    },
-    multiselect: async (message, options) => {
-      multiselectCalls.push({ message, options })
-      const answer = nextAnswer(message)
-      if (!Array.isArray(answer)) {
-        throw new Error(
-          `multiselect needs a string[] scripted answer, got: ${String(answer)}`,
-        )
-      }
-      return answer
-    },
-    text: async (message, options) => {
-      textCalls.push({ message, defaultValue: options?.defaultValue })
-      const answer = String(nextAnswer(message))
-      // Mirrors @clack/prompts: an empty submission resolves to defaultValue.
-      if (answer === "" && options?.defaultValue !== undefined)
-        return options.defaultValue
-      return answer
-    },
-    password: async (message) => String(nextAnswer(message)),
-    confirm: async (message, initialValue) => {
-      confirmCalls.push({ message, initialValue })
-      return Boolean(nextAnswer(message))
-    },
-    spinner: () => ({ start: () => {}, stop: () => {} }),
-  }
-
-  return {
-    prompts,
-    asked,
-    errors,
-    multiselectCalls,
-    confirmCalls,
-    selectCalls,
-    textCalls,
-  }
-}
+import { createScriptedPrompts } from "./command-stubs.js"
 
 describe("readOptionalValue", () => {
   it("reads the value of an active line", () => {
