+++ cli/tsconfig.json
@@ -14,5 +14,5 @@
     "isolatedModules": true
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"]
+  "exclude": ["src/**/__tests__/**"]
 }
